fix(ci): resolve develop-branch build + test failures
Four independent root causes, one PR — all inherited by every PR stacked
off develop. After this lands, CI goes green on develop and on all 10
ts-sdk-perf PRs (#56-65) once they rebase.

## 1. `/vaults` Suspense boundary (Next.js 16 prerender)

`app/vaults/VaultShell.tsx` calls `useSearchParams()` to parse `?vault=id`
deeplinks. Next.js 16 requires such calls to be inside a `<Suspense>`
boundary during static prerender, or the build aborts with
`useSearchParams() should be wrapped in a suspense boundary`.

Wrapped `<VaultShell />` in `<Suspense fallback={null}>` at the page root.
Invisible to users — short-circuits during prerender, hydrates on mount.

## 2. `/wallet` Suspense boundary

Same root cause at `app/wallet/page.tsx` which used `useSearchParams()`
for the `?tab=` deeplink. The page was a large `'use client'` component,
so wrapping it in Suspense required splitting:
- `app/wallet/page.tsx` → tiny Suspense wrapper (server-safe)
- `app/wallet/WalletDashboardContent.tsx` → new file with all the
  original dashboard logic (git rename; 258 lines preserved)

## 3. Setup.ts race — `ecc library invalid` / hang-on-SDK-import

`__tests__/setup.ts` initialized the SDK via a fire-and-forget
`.then(...)`. Under parallel Vitest pool load in CI, tests that imported
the SDK on their own (`buildAlkaneTransferPsbt`, `usePools`) could race
the setup, hanging until their 30s timeout fired.

Switched to `await import(...)` at top level so every test starts with a
fully-initialized SDK. Also downgraded the catch log to a `console.warn`
since the bundled bip32's testEcc check fails in some Node CI
environments even though the SDK is otherwise usable.

## 4. Stale `.claude/worktrees/*` snapshots running as tests

The `.claude/worktrees/` directory contains frozen snapshots of prior
agent sessions. Their test files pinned regex assertions against source
that has since changed, producing confusing pseudo-failures unrelated
to the current codebase.

Added `.claude/**` to `vitest.config.ts` excludes. Dropped unit-test
file count from ~418 to ~57, all in live app code.

## 5. wallet-dashboard.test.ts assertion drift

The `destructures balances, isLoading, error, refresh` test hardcoded
a regex that expected plain `{ balances, isLoading, ... }`. The actual
source aliases via `{ balances, isAlkaneLoading: isLoading, ... }`.
Updated the regex to accept either form. Also pointed the three
`page.tsx` source-reads at `WalletDashboardContent.tsx` (the split target).

## Verification

- `tsc --noEmit` exits 0
- `pnpm run test:unit`: 57 files / 1,936 tests / 0 failures / 142 skipped
- `pnpm run build`: succeeds, `/vaults` and `/wallet` both prerender as
  static content
